Outsurce button function in button.js
