+++ src/components/button/button.js
@@ -1,0 +1,7 @@
+export function createButton(buttonText) {
+  const button = document.createElement("button");
+  button.classList.add("btn");
+  const text = document.createTextNode(buttonText);
+  button.append(text);
+  return button;
+}
